requirements.txt
argon2-cffi
bcrypt
pbkdf2
scrypt
